rollup file cleanup
